+++ mod.ts
@@ -1,6 +1,33 @@
-export * from "./unknown/mod.ts";
-export { type Primitive } from "./types.ts";
-export * from "./number/mod.ts";
-export * from "./string/mod.ts";
-export * from "./date/mod.ts";
+export {
+  isAsyncIterable,
+  isBigint,
+  isBoolean,
+  isDate,
+  isError,
+  isFalse,
+  isFalsy,
+  isFunction,
+  isIterable,
+  isNull,
+  isNullable,
+  isNumber,
+  isObject,
+  isPrimitive,
+  isPromise,
+  isString,
+  isSymbol,
+  isTrue,
+  isTruthy,
+  isUndefined,
+  type Primitive,
+} from "./top_types.ts";
+export {
+  isEven,
+  isNegativeNumber,
+  isNonNegativeInteger,
+  isOdd,
+  isPositiveNumber,
+} from "./numbers.ts";
+export { isHexColorFormat } from "./strings.ts";
+export { isValidDate } from "./dates.ts";
 export { hasOwn } from "./object.ts";