+++ types.ts
@@ -1,9 +1,0 @@
-/** Types for Primitive value. */
-export type Primitive =
-  | number
-  | string
-  | boolean
-  | bigint
-  | undefined
-  | null
-  | symbol;